Reconfigured dockerfiles and Jenkinsfile for rocRAND
[ROCm/rocRAND commit: 69406cf4e6e70acb8b51c44e0e0b8eff981a1656]

+++ Jenkinsfile
@@ -6,12 +6,13 @@ import com.amd.docker.*
 ////////////////////////////////////////////////////////////////////////
 import java.nio.file.Path;
 
-rocprimCI:
+rocRANDCI:
 {
 
-    def rocprim = new rocProject('rocprim')
+    def rocrand = new rocProject('rocRAND')
 
-    def nodes = new dockerNodes(['gfx803', 'gfx900', 'gfx906'], rocprim)
+    def nodes = new dockerNodes(['gfx900 && ubuntu', 'gfx906 && ubuntu', 'gfx900 && centos7', 'gfx900 && ubuntu && hip-clang',
+                    'gfx906 && ubuntu && hip-clang'], rocrand)
 
     boolean formatCheck = false
      
@@ -20,11 +21,26 @@ rocprimCI:
         platform, project->
 
         project.paths.construct_build_prefix()
-        def command = """#!/usr/bin/env bash
-                  set -x
-                  cd ${project.paths.project_build_prefix}
-                  LD_LIBRARY_PATH=/opt/rocm/hcc/lib CXX=${project.compiler.compiler_path} ${project.paths.build_command}
-                """
+        
+        def command 
+
+        if(platform.jenkinsLabel.contains('hip-clang'))
+        {
+            command = """#!/usr/bin/env bash
+                    set -x
+                    cd ${project.paths.project_build_prefix}
+                    LD_LIBRARY_PATH=/opt/rocm/hcc/lib CXX=/opt/rocm/bin/hipcc ${project.paths.build_command} --hip-clang
+                    """
+        }
+        else
+        {
+            command = """#!/usr/bin/env bash
+                    set -x
+                    cd ${project.paths.project_build_prefix}
+                    LD_LIBRARY_PATH=/opt/rocm/hcc/lib CXX=/opt/rocm/bin/hcc ${project.paths.build_command}
+                    """
+        }
+
         platform.runCommand(this, command)
     }
 
@@ -32,14 +48,26 @@ rocprimCI:
     {
         platform, project->
 
-        def testCommand = 'ctest --output-on-failure -E rocprim.hip.device_merge_sort'
+        def command
 
-        def command = """#!/usr/bin/env bash
-                        set -x
-                        cd ${project.paths.project_build_prefix}
-                        cd ${project.testDirectory}
-                        LD_LIBRARY_PATH=/opt/rocm/hcc/lib ${testCommand}
-                  """
+        if(platform.jenkinsLabel.contains('centos'))
+        {
+            command = """#!/usr/bin/env bash
+                    set -x
+                    cd ${project.paths.project_build_prefix}/build/release
+                    make -j4
+                    sudo ctest --output-on-failure
+                """
+        }
+        else
+        {
+            command = """#!/usr/bin/env bash
+                    set -x
+                    cd ${project.paths.project_build_prefix}/build/release
+                    make -j4
+                    ctest --output-on-failure
+                """
+        }
 
         platform.runCommand(this, command)
     }
@@ -48,21 +76,42 @@ rocprimCI:
     {
         platform, project->
 
-        def command = """
-                      set -x
-                      cd ${project.paths.project_build_prefix}/build
-                      make package
-                      rm -rf package && mkdir -p package
-                      mv *.deb package/
-                      dpkg -c package/*.deb
-                      """
-
-        platform.runCommand(this, command)
-        platform.archiveArtifacts(this, """${project.paths.project_build_prefix}/build/package/*.deb""")
+        def command
+        
+        if(platform.jenkinsLabel.contains('centos'))
+        {
+            command = """
+                    set -x
+                    cd ${project.paths.project_build_prefix}/build/release
+                    make package
+                    rm -rf package && mkdir -p package
+                    mv *.rpm package/
+                    rpm -qlp package/*.rpm
+                  """
+            
+            platform.runCommand(this, command)
+            platform.archiveArtifacts(this, """${project.paths.project_build_prefix}/build/release/package/*.rpm""")
+        }
+        else if(platform.jenkinsLabel.contains('hip-clang'))
+        {
+            packageCommand = null
+        }
+        else
+        {
+            command = """
+                    set -x
+                    cd ${project.paths.project_build_prefix}/build/release
+                    make package
+                    rm -rf package && mkdir -p package
+                    mv *.deb package/
+                    dpkg -c package/*.deb
+                  """        
+            
+            platform.runCommand(this, command)
+            platform.archiveArtifacts(this, """${project.paths.project_build_prefix}/build/release/package/*.deb""")
+        }
     }
 
-    buildProject(rocprim, formatCheck, nodes.dockerArray, compileCommand, testCommand, packageCommand)
-
+    buildProject(rocrand, formatCheck, nodes.dockerArray, compileCommand, testCommand, packageCommand)
 }
-
